Modernize wrappers for Schema and related classes.
Covers AliasMap, Schema, and SchemaMapper.
