+++ python/lsst/afw/table/baseColumnView/baseColumnView.cc
@@ -118,8 +118,6 @@ static void declareBitsColumn(py::module &mod) {
 }
 
 PYBIND11_MODULE(baseColumnView, mod) {
-    py::module::import("lsst.afw.table.schema");
-
     declareBaseColumnView(mod);
     declareBitsColumn(mod);
 }
