+++ python/lsst/afw/table/multiMatch.py
@@ -23,7 +23,7 @@ import collections.abc
 import numpy
 
 import lsst.geom
-from .schemaMapper import SchemaMapper
+from ._schemaMapper import SchemaMapper
 from .aggregates import CoordKey
 from .source import SourceRecord
